backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
deprecation==2.1.0
docopt==0.6.2
fastapi==0.121.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==25.0
pipreqs==0.4.13
postgrest==2.24.0
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
realtime==2.24.0
requests==2.32.5
sniffio==1.3.1
starlette==0.49.3
storage3==2.24.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.22
uvicorn==0.34.3
websockets==15.0.1
yarg==0.1.10
yarl==1.22.0
gTTS==2.5.4
Pillow==11.0.0
faster-whisper==1.0.3  # Оптимизированная версия Whisper (меньше памяти, быстрее)
ffmpeg-python==0.2.0  # Python wrapper для ffmpeg

